Cloddy sucks
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C837

+++ inc/Game/GameState.h
@@ -22,7 +22,7 @@ namespace TikiEngine
 			GameState(Engine* engine, SceneLevel* scene);
 			~GameState();
 
-			bool LoadLevel(UInt32 id);
+			bool LoadLevel(Int64 id);
 			void DisposeLevel();
 
 			Engine* GetEngine();
